Feat : Comment (hospital and doctor)

+++ Models/DoctorComment.cs
@@ -6,8 +6,8 @@ namespace Models
     {
         public int DoctorCommentId { get; set; }
         public DateTime? Time { get; set; }
-        public int DoctorId { get; set; }
-        public int PatientId { get; set; }
+        public long DoctorId { get; set; }
+        public long PatientId { get; set; }
         public string? Comment { get; set; }
         public float? Point { get; set; }
